[cleanup] Adapted the assert comparing with std::size_t

+++ bindings/python/crocoddyl/core/activation-base.hpp
@@ -22,13 +22,13 @@ class ActivationModelAbstract_wrap : public ActivationModelAbstract, public bp::
       : ActivationModelAbstract(nr), bp::wrapper<ActivationModelAbstract>() {}
 
   void calc(const boost::shared_ptr<ActivationDataAbstract>& data, const Eigen::Ref<const Eigen::VectorXd>& r) {
-    assert(r.size() == nr_ && "r has wrong dimension");
+    assert(static_cast<std::size_t>(r.size()) == nr_ && "r has wrong dimension");
     return bp::call<void>(this->get_override("calc").ptr(), data, (Eigen::VectorXd)r);
   }
 
   void calcDiff(const boost::shared_ptr<ActivationDataAbstract>& data, const Eigen::Ref<const Eigen::VectorXd>& r,
                 const bool& recalc = true) {
-    assert(r.size() == nr_ && "r has wrong dimension");
+    assert(static_cast<std::size_t>(r.size()) == nr_ && "r has wrong dimension");
     return bp::call<void>(this->get_override("calcDiff").ptr(), data, (Eigen::VectorXd)r, recalc);
   }
 };
