+++ bindings/python/crocoddyl/multibody/cost-base.hpp
@@ -33,15 +33,15 @@ class CostModelAbstract_wrap : public CostModelAbstract, public bp::wrapper<Cost
 
   void calc(const boost::shared_ptr<CostDataAbstract>& data, const Eigen::Ref<const Eigen::VectorXd>& x,
             const Eigen::Ref<const Eigen::VectorXd>& u) {
-    assert(x.size() == state_.get_nx() && "x has wrong dimension");
-    assert((u.size() == nu_ || nu_ == 0) && "u has wrong dimension");
+    assert(static_cast<std::size_t>(x.size()) == state_.get_nx() && "x has wrong dimension");
+    assert((static_cast<std::size_t>(u.size()) == nu_ || nu_ == 0) && "u has wrong dimension");
     return bp::call<void>(this->get_override("calc").ptr(), data, (Eigen::VectorXd)x, (Eigen::VectorXd)u);
   }
 
   void calcDiff(const boost::shared_ptr<CostDataAbstract>& data, const Eigen::Ref<const Eigen::VectorXd>& x,
                 const Eigen::Ref<const Eigen::VectorXd>& u, const bool& recalc = true) {
-    assert(x.size() == state_.get_nx() && "x has wrong dimension");
-    assert((u.size() == nu_ || nu_ == 0) && "u has wrong dimension");
+    assert(static_cast<std::size_t>(x.size()) == state_.get_nx() && "x has wrong dimension");
+    assert((static_cast<std::size_t>(u.size()) == nu_ || nu_ == 0) && "u has wrong dimension");
     return bp::call<void>(this->get_override("calcDiff").ptr(), data, (Eigen::VectorXd)x, (Eigen::VectorXd)u, recalc);
   }
 };
