+++ src/core/activations/quadratic.cpp
@@ -16,13 +16,13 @@ ActivationModelQuad::~ActivationModelQuad() {}
 
 void ActivationModelQuad::calc(const boost::shared_ptr<ActivationDataAbstract>& data,
                                const Eigen::Ref<const Eigen::VectorXd>& r) {
-  assert(r.size() == nr_ && "r has wrong dimension");
+  assert(static_cast<std::size_t>(r.size()) == nr_ && "r has wrong dimension");
   data->a_value = 0.5 * r.transpose() * r;
 }
 
 void ActivationModelQuad::calcDiff(const boost::shared_ptr<ActivationDataAbstract>& data,
                                    const Eigen::Ref<const Eigen::VectorXd>& r, const bool& recalc) {
-  assert(r.size() == nr_ && "r has wrong dimension");
+  assert(static_cast<std::size_t>(r.size()) == nr_ && "r has wrong dimension");
   if (recalc) {
     calc(data, r);
   }
