+++ src/multibody/actions/impulse-fwddyn.cpp
@@ -34,7 +34,7 @@ ActionModelImpulseFwdDynamics::~ActionModelImpulseFwdDynamics() {}
 void ActionModelImpulseFwdDynamics::calc(const boost::shared_ptr<ActionDataAbstract>& data,
                                          const Eigen::Ref<const Eigen::VectorXd>& x,
                                          const Eigen::Ref<const Eigen::VectorXd>& u) {
-  assert(x.size() == state_.get_nx() && "x has wrong dimension");
+  assert(static_cast<std::size_t>(x.size()) == state_.get_nx() && "x has wrong dimension");
 
   const std::size_t& nq = state_.get_nq();
   const std::size_t& nv = state_.get_nv();
@@ -73,7 +73,7 @@ void ActionModelImpulseFwdDynamics::calc(const boost::shared_ptr<ActionDataAbstr
 void ActionModelImpulseFwdDynamics::calcDiff(const boost::shared_ptr<ActionDataAbstract>& data,
                                              const Eigen::Ref<const Eigen::VectorXd>& x,
                                              const Eigen::Ref<const Eigen::VectorXd>& u, const bool& recalc) {
-  assert(x.size() == state_.get_nx() && "x has wrong dimension");
+  assert(static_cast<std::size_t>(x.size()) == state_.get_nx() && "x has wrong dimension");
 
   const std::size_t& nv = state_.get_nv();
   const std::size_t& ni = impulses_.get_ni();
