+++ src/multibody/actuations/floating-base.cpp
@@ -25,7 +25,7 @@ ActuationModelFloatingBase::~ActuationModelFloatingBase() {}
 void ActuationModelFloatingBase::calc(const boost::shared_ptr<ActuationDataAbstract>& data,
                                       const Eigen::Ref<const Eigen::VectorXd>&,
                                       const Eigen::Ref<const Eigen::VectorXd>& u) {
-  assert(u.size() == nu_ && "u has wrong dimension");
+  assert(static_cast<std::size_t>(u.size()) == nu_ && "u has wrong dimension");
   data->tau.tail(nu_) = u;
 }
